Extracts icon config and exporting
