+++ src/common/types.ts
@@ -64,3 +64,10 @@ export interface AirtableData {
   storage: Option[];
   accountable: Option[];
 }
+
+export interface IconConfig {
+  hexSrc: string;
+  iconSrc: string;
+  fontColor: string;
+  iconName: string;
+}
